move ExecOps mounting logic into new package so logic can be reused from gateway exec
